fix: use Decimal for currency handling, migrated from float (#51)
* fix: use Decimal for currency handling, migrated from float

* fix: use Decimal for currency handling, migrated from float

+++ backend/app/backtesting/actions.py
@@ -1,16 +1,17 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
 class BuyAction:
     asset_id: int
-    dollar_amount: float
+    dollar_amount: Decimal
 
 
 @dataclass
 class SellAction:
     asset_id: int
-    quantity: float
+    quantity: Decimal
 
 
 Action = BuyAction | SellAction
